Title: 범위 내에 있는 점의 수 2, Time: 219ms, Memory: 8MB, Status: Wrong Answer - CodeTree

+++ 240913/범위 내에 있는 점의 수 2/the-number-of-points-within-the-range-2.cpp
@@ -2,8 +2,8 @@
 #include <algorithm>
 using namespace std;
 
-long a[2000003];
-long sum[2000003];
+int a[2000003];
+int sum[2000003];
 int main() 
 {
     int n, q;
